Added possibility to delete haiku
- improved php delete script
- fixed bugs in haiku operations because of DB structure
changes
- now post menu changes while user is logged in

+++ resources/haiku_delete.php
@@ -10,17 +10,47 @@
         require_once "../config/config.php";
         require_once "db_connect.php";
 
+        $check = $conn->prepare("SELECT background, handwriting FROM haiku WHERE id = :hid");
+
+        try
+        {
+            $check->bindParam(":hid", $_POST['haiku_id']);
+            $check->execute();
+        }
+        catch(Exception $e)
+        {
+            die(json_encode([false, "Error, cannot delete haiku, try later!"]));
+        }
+
+        if($check->rowCount() == 0)
+        {
+            die(json_encode([false, "Error, cannot find haiku to delete!"]));
+        }
+
         $query = $conn->prepare("DELETE FROM haiku WHERE id = :hid");
-        $query->bindParam(":hid", $_POST['haiku_id']);
+        $query2 = $conn->prepare("UPDATE haiku_reports SET solved = 1 WHERE haiku_id = :hid");
         
         try
         {
+            $query->bindParam(":hid", $_POST['haiku_id']);
+            $query2->bindParam(":hid", $_POST['haiku_id']);
             $query->execute();
+            $query2->execute();
             $result = array(true, "Haiku deleted successfully.");
         }
         catch(Exception $e)
         {
-            $result = array(false, "Error, cannot delete haiku, try later!");
+            die(json_encode([false, "Error, cannot delete haiku, try later!"]));
+        }
+
+        $images = $check->fetchAll();
+
+        foreach($images as $image)
+        {
+            if(file_exists(HW_DIR . $image['handwriting']) && $image['handwriting'] != "no_hw.jpg")
+                unlink(HW_DIR . $image['handwriting']);
+            else if(file_exists(BG_DIR . $image['background']) && $image['background'] != "default.png")
+                unlink(BG_DIR . $image['background']);
         }
     }
     echo json_encode($result);
